CF-3098
(cherry picked from commit 82062547bbb438cd987784db123bab1de28f04da)

+++ i21Database/dbo/Stored Procedures/uspCFInsertTransactionRecord.sql
@@ -1358,7 +1358,7 @@ BEGIN
 		DECLARE @intPrcItemUOMId			INT
 		DECLARE @dblPrcPriceOut				NUMERIC(18,6)	
 		DECLARE @strPrcPricingOut			NVARCHAR(MAX)		
-		DECLARE @intPrcAvailableQuantity	INT				
+		DECLARE @intPrcAvailableQuantity	NUMERIC(18,6)				
 		DECLARE @dblPrcOriginalPrice		NUMERIC(18,6)	
 		DECLARE @intPrcContractHeaderId		INT				
 		DECLARE @intPrcContractDetailId		INT
